+++ src/Queries.Core/Parts/Columns/ColumnBase.cs
@@ -17,19 +17,38 @@ public abstract class ColumnBase : IColumn, IEquatable<ColumnBase>
     ///<inheritdoc/>
     public abstract bool Equals(ColumnBase other);
 
-    ///<inheritdoc/>
+    /// <summary>
+    /// Provides the ability to define type conversion operators for custom casting to <see cref="ColumnBase"/>.
+    /// </summary>
     public static implicit operator ColumnBase(int value) => new NumericColumn(value);
 
-    ///<inheritdoc/>
+    /// <summary>
+    /// Allows implicit conversion from double to <see cref="ColumnBase"/>.
+    /// </summary>
+    /// <param name="value">The double value to convert.</param>
+    /// <returns>
+    /// A new instance of <see cref="NumericColumn"/> representing the double value.
+    /// </returns>
     public static implicit operator ColumnBase(double value) => new NumericColumn(value);
 
-    ///<inheritdoc/>
+    /// <summary>
+    /// Defines the implicit cast operator from the <see cref="float"/> type to a <see cref="ColumnBase"/> object.
+    /// </summary>
+    /// <param name="value">The floating-point value to cast.</param>
+    /// <returns>An instance of <see cref="NumericColumn"/> representing the given value.</returns>
     public static implicit operator ColumnBase(float value) => new NumericColumn(value);
 
-    ///<inheritdoc/>
+    /// <summary>
+    /// Defines an operator that enables implicit casting to <see cref="ColumnBase"/> from various types such as
+    /// <see cref="int"/>, <see cref="string"/>, <see cref="bool"/>, <see cref="double"/>, <see cref="float"/>,
+    /// <see cref="DateTime"/>, <see cref="DateOnly"/>, and <see cref="TimeOnly"/>.
+    /// </summary>
     public static implicit operator ColumnBase(string value) => new StringColumn(value);
 
-    ///<inheritdoc/>
+    /// <summary>
+    /// Defines a custom implicit conversion operator for a specific source or target type.
+    /// Enables seamless type conversion between primitive types and <see cref="ColumnBase"/> or its derived types.
+    /// </summary>
     public static implicit operator ColumnBase(bool value) => new BooleanColumn(value);
 
     ///<inheritdoc/>